splat: added typecasting

+++ splat/src/main/uk/ac/starlink/splat/vo/TapQueryWithDatalink.java
@@ -66,7 +66,7 @@ public class TapQueryWithDatalink extends TapQuery {
             StringBuffer sbuf = new StringBuffer()
                 .append( "TAP response is not a VOTable" );
             byte[] buf = in.getHeadBuffer();
-            int nb = Math.min( in.getReadCount(), buf.length );
+            int nb = Math.min( (int) in.getReadCount(), (int) buf.length );
             if ( nb > 0 ) {           
                 sbuf.append( " - " ).append( new String( buf, 0, nb, "UTF-8" ) );
                 if ( nb == buf.length ) {
